Remove obsolete Vectors class to reduce confusion
